// speedyGo-ui/src/app/services/models/post.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Comment } from '../models/comment';
import { Reaction } from '../models/reaction';
import { Sharing } from '../models/sharing';
import { User } from '../models/user';
export interface Post {
  comments?: Array<Comment>;
  content?: string;
  createdDate?: string;
  id?: number;
  lastModifiedDate?: string;
  mediaUrl?: string;
  reactions?: Array<Reaction>;
  senderPost?: User;
  sharingList?: Array<Sharing>;
  status?: 'ACTIF' | 'DELETED' | 'ARCHIVED';
  title?: string;
  visibility?: boolean;
  sentimentPictureScore?:boolean;
  sentimentScore?: string;
}