// speedyGo-ui/src/app/services/models/post-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Comment } from '../models/comment';
import { Reaction } from '../models/reaction';
import {Sharing} from "./sharing";
export interface PostResponse {
  comments?: Array<Comment>;
  content?: string;
  createdDate?: string;
  id?: number;
  lastModifiedDate?: string;
  mediaUrl?: string;
  reactions?: Array<Reaction>;
  status?: 'ACTIF' | 'DELETED' | 'ARCHIVED';
  title?: string;
  visibility?: boolean;
  sharingList?:Array<Sharing>;
  sentimentPictureScore?:boolean;
  sentimentScore?: string;
}