// speedyGo-ui/src/app/services/models/page-response-post-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PostResponse } from '../models/post-response';
export interface PageResponsePostResponse {
  content?: Array<PostResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}